Email a notification to managers upon termination

+++ php/orm/Site.php
@@ -597,7 +597,7 @@ class Site
 				mysqli_query($dbconn, "INSERT INTO SiteManager (`UserFKOfManager`, `SiteFK`, `Approved`) VALUES ('" . $manager->getID() . "', '" . $this->id . "', '0')");
 				//$id = intval(mysqli_insert_id($dbconn));
 				$message = "<div style=\"text-align:center;border-radius:5px;padding:20px;font-family:'Segoe UI', Frutiger, 'Frutiger Linotype', 'Dejavu Sans', 'Helvetica Neue', Arial, sans-serif;\"><div style=\"text-align:left;color:#777;margin-bottom:40px;font-size:20px;\">" . $this->getCreator()->getFullName() . " would like you to be a manager of the \"" . $this->getName() . "\" site in " . $this->getRegion() . ". Please sign in to <a href='https://caterpillarscount.unc.edu/signIn'>caterpillarscount.unc.edu</a> using this email address (" . $manager->getEmail() . ") to approve or deny this request.</div><a href='https://caterpillarscount.unc.edu/signIn'><button style=\"border:0px none transparent;background:#fed136; border-radius:5px;padding:20px 40px;font-size:20px;color:#fff;font-family:'Segoe UI', Frutiger, 'Frutiger Linotype', 'Dejavu Sans', 'Helvetica Neue', Arial, sans-serif;font-weight:bold;cursor:pointer;\">SIGN IN NOW</button></a><div style=\"padding-top:40px;margin-top:40px;margin-left:-40px;margin-right:-40px;border-top:1px solid #eee;color:#bbb;font-size:14px;\"></div></div>";
-				email($manager->getEmail(), "New Caterpillars Count! Site Manager Request!", $message);
+				email($manager->getEmail(), "New Caterpillars Count! site manager request!", $message);
 			}
 			mysqli_close($dbconn);
 			return true;
@@ -608,7 +608,17 @@ class Site
 	public function terminateManager($manager){
 		if(is_object($manager) && get_class($manager) == "User" && $manager->getID() != $this->creator->getID()){
 			$dbconn = (new Keychain)->getDatabaseConnection();
-			$query = mysqli_query($dbconn, "DELETE FROM `SiteManager` WHERE `UserFKOfManager`='" . $manager->getID() . "' AND `SiteFK`='" . $this->id . "'");
+			$query = mysqli_query($dbconn, "SELECT `Approved` FROM `SiteManager` WHERE `UserFKOfManager`='" . $manager->getID() . "' AND `SiteFK`='" . $this->id . "' LIMIT 1");
+			if(mysqli_num_rows($query) == 0){
+				$approved = filter_var(mysqli_fetch_assoc($query)["Approved"], FILTER_VALIDATE_BOOLEAN);
+				$query = mysqli_query($dbconn, "DELETE FROM `SiteManager` WHERE `UserFKOfManager`='" . $manager->getID() . "' AND `SiteFK`='" . $this->id . "'");
+				$approvedMessageAddOn = "";
+				if($approved){
+					$approvedMessageAddOn = " Thank you for the time you've dedicated to managing this site in the past.";
+				}
+				$message = "<div style=\"text-align:center;border-radius:5px;padding:20px;font-family:'Segoe UI', Frutiger, 'Frutiger Linotype', 'Dejavu Sans', 'Helvetica Neue', Arial, sans-serif;\"><div style=\"text-align:left;color:#777;margin-bottom:40px;font-size:20px;\">" . $this->getCreator()->getFullName() . " no longer requires your services as a manager of the \"" . $this->getName() . "\" site in " . $this->getRegion() . "." . $approvedMessageAddOn . "</div><div style=\"padding-top:40px;margin-top:40px;margin-left:-40px;margin-right:-40px;border-top:1px solid #eee;color:#bbb;font-size:14px;\"></div></div>";
+				email($manager->getEmail(), "Your Caterpillars Count! \"" . $this->getName() . "\" managment services are no longer required.", $message);
+			}
 			mysqli_close($dbconn);
 			return true;
 		}
